+++ packages/ui/package.json
@@ -7,17 +7,17 @@
     "#*": "./src/*"
   },
   "exports": {
-    "./ArticleHeader": "./src/ArticleHeader.tsx",
+    "./Article": "./src/Article/index.ts",
+    "./Article/*": "./src/Article/*.tsx",
     "./Avatar": "./src/Avatar.tsx",
     "./AvatarGroup": "./src/AvatarGroup.tsx",
     "./BaseLink": "./src/BaseLink.tsx",
     "./Button": "./src/Button.tsx",
+    "./BlogPostHeader": "./src/BlogPostHeader.tsx",
     "./BreadCrumbs": "./src/BreadCrumbs.tsx",
     "./BreadCrumbsLayout": "./src/BreadCrumbsLayout.tsx",
     "./BreakpointDebugger": "./src/BreakpointDebugger.tsx",
     "./DescriptionText": "./src/DescriptionText.tsx",
-    "./DigestArticle": "./src/DigestArticle/index.ts",
-    "./DigestArticle/*": "./src/DigestArticle/*.tsx",
     "./DigestArticleHeader": "./src/DigestArticleHeader.tsx",
     "./ErrorMessage": "./src/ErrorMessage.tsx",
     "./GlobalError": "./src/GlobalError.tsx",
