--- packages/ui/src/DigestArticle/DigestArticleLayout.tsx
@@ -1,7 +1,0 @@
-type DigestArticleLayoutProps = {
-  children: React.ReactNode
-}
-
-export function DigestArticleLayout({ children }: DigestArticleLayoutProps) {
-  return <div className="m-auto max-w-2xl space-y-16">{children}</div>
-}
