--- packages/ui/src/DigestArticle/index.ts
@@ -1,2 +1,0 @@
-export { DigestArticleLayout } from './DigestArticleLayout'
-export { DigestArticleHeader } from './DigestArticleHeader'
